Actualizado Librerias | Livewire 3 | Jecstream

+++ app/Livewire/Dashboard/ParametrosComponent.php
@@ -1,31 +1,52 @@
 <?php
 
-namespace App\Http\Livewire\Dashboard;
+namespace App\Livewire\Dashboard;
 
 use App\Models\Parametro;
 use Illuminate\Validation\Rule;
 use Jantinnerezo\LivewireAlert\LivewireAlert;
+use Livewire\Attributes\On;
 use Livewire\Component;
-use Livewire\WithPagination;
 
 class ParametrosComponent extends Component
 {
     use LivewireAlert;
-    use WithPagination;
 
-    protected $paginationTheme = 'bootstrap';
-    protected $listeners = ['buscar', 'confirmed'];
-
+    public $rows = 0, $numero = 14, $tableStyle = false;
     public $view = "create", $keyword;
     public $parametro_id, $nombre, $tabla_id, $valor;
 
+    public function mount()
+    {
+        $this->setLimit();
+    }
+
     public function render()
     {
-        $parametros = Parametro::buscar($this->keyword)->orderBy('updated_at', 'ASC')->paginate(numRowsPaginate());
+        $parametros = Parametro::buscar($this->keyword)
+            ->orderBy('updated_at', 'ASC')
+            ->limit($this->rows)
+            ->get();
+
         $rows = Parametro::count();
+
+        if ($rows > $this->numero) {
+            $this->tableStyle = true;
+        }
+
         return view('livewire.dashboard.parametros-component')
             ->with('parametros', $parametros)
-            ->with('rows', $rows);
+            ->with('rowsParametros', $rows);
+    }
+
+    public function setLimit()
+    {
+        if (numRowsPaginate() < $this->numero) {
+            $rows = $this->numero;
+        } else {
+            $rows = numRowsPaginate();
+        }
+        $this->rows = $this->rows + $rows;
     }
 
     public function limpiar()
@@ -33,6 +54,7 @@ class ParametrosComponent extends Component
         $this->reset([
             'parametro_id', 'nombre', 'tabla_id', 'valor', 'view', 'keyword'
         ]);
+        $this->resetErrorBag();
     }
 
     protected function rules($id = null)
@@ -61,29 +83,34 @@ class ParametrosComponent extends Component
             $message = "Parametro Actualizado";
         }
 
-        $parametro->nombre = $this->nombre;
-        if (!empty($this->tabla_id)){
-            $parametro->tabla_id = $this->tabla_id;
-        }
-        if (!empty($this->valor)){
-            $parametro->valor = $this->valor;
-        }
-        $parametro->save();
+        if ($parametro){
+            $parametro->nombre = $this->nombre;
+            if (!empty($this->tabla_id)){
+                $parametro->tabla_id = $this->tabla_id;
+            }
+            if (!empty($this->valor)){
+                $parametro->valor = $this->valor;
+            }
+            $parametro->save();
 
-        $this->alert($type, $message);
+            $this->alert($type, $message);
+        }
         $this->limpiar();
     }
 
     public function edit($id)
     {
         $parametro = Parametro::find($id);
-        $this->parametro_id = $parametro->id;
-        $this->nombre = $parametro->nombre;
-        $this->tabla_id = $parametro->tabla_id;
-        $this->valor = $parametro->valor;
-        $this->view = "edit";
+        if ($parametro){
+            $this->parametro_id = $parametro->id;
+            $this->nombre = $parametro->nombre;
+            $this->tabla_id = $parametro->tabla_id;
+            $this->valor = $parametro->valor;
+            $this->view = "edit";
+        }
     }
 
+    #[On('buscar')]
     public function buscar($keyword)
     {
         $this->keyword = $keyword;
@@ -103,15 +130,15 @@ class ParametrosComponent extends Component
         ]);
     }
 
+    #[On('confirmed')]
     public function confirmed()
     {
         $parametro = Parametro::find($this->parametro_id);
-        $parametro->delete();
-        $this->limpiar();
-        $this->alert(
-            'success',
-            'Parametro Eliminado'
-        );
+        if ($parametro){
+            $parametro->delete();
+            $this->limpiar();
+            $this->alert('success', 'Parametro Eliminado.');
+        }
     }
 
 }
